+++ bpa/stor/src/bplib_stor_sql_store.c
@@ -178,7 +178,7 @@ SQL_Status_t BPLib_SQL_StoreChunk(int64_t BundleRowID, const void* Chunk, size_t
 SQL_Status_t BPLib_SQL_StoreBundle(sqlite3* db, BPLib_Bundle_t* Bundle, BPLib_BundleCache_t* BundleCache)
 {
     SQL_Status_t       SQLStatus;
-    uint8_t            BundleRowID;
+    uint64_t           BundleRowID;
     BPLib_MEM_Block_t* CurrMemBlock;
 
     /* Store the indexable metadata */
@@ -220,7 +220,7 @@ SQL_Status_t BPLib_SQL_StoreImpl(BPLib_Instance_t* Inst, size_t *TotalBytesStore
                                 size_t *DuplicateBundlesIgnored)
 {
     SQL_Status_t SQLStatus;
-    uint8_t      i;
+    size_t       i;
     sqlite3*     db;
     size_t       NewBundleBytes;
